fix: go builtin types to egoist.types
