+++ examples/e2e/generate/04use-di/internal.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 import typing as t
-from prestring.go.codeobject import Module, Symbol
+from egoist.types import GoError
 
 
 class Message:
@@ -13,25 +13,6 @@ class Greeter:
 
 class Event:
     gopackage = "m/internal"
-
-
-class GoError:
-    name = "err"
-    priority = 10
-
-    @classmethod
-    def emit(self, m: Module, err: Symbol) -> None:
-        with m.if_(f"{err} != nil"):
-            m.return_(err)
-
-
-class GoTeardown:
-    name = "teardown"
-    priority = 1
-
-    @classmethod
-    def emit(self, m: Module, teardown: Symbol) -> None:
-        m.stmt(f"defer {teardown}()")
 
 
 def NewMessage() -> Message:
